Ready for aws
